game/client: Adjust types for color <-> float conversions and fix some float cast warnings

+++ game/client/c_baseplayer.cpp
@@ -1382,7 +1382,7 @@ void C_BasePlayer::CreateWaterEffects( void )
 		pParticle->m_vecVelocity = RandomVector( -2.0f, 2.0f );
 
 		//FIXME: We should tint these based on the water's fog value!
-		int color = random->RandomInt( 32, 128 );
+		byte color = static_cast<byte>(random->RandomInt( 32, 128 ));
 		pParticle->m_uchColor[0] = color;
 		pParticle->m_uchColor[1] = color;
 		pParticle->m_uchColor[2] = color;
